moved some things out into their own components and some state into a hook
