+++ src/components/quickstart/index.ts
@@ -1,0 +1,5 @@
+export { AssetsPanel } from './AssetsPanel'
+export { ColorPicker } from './ColorPicker'
+export { ColorSettingsPanel } from './ColorSettingsPanel'
+export { TemplateContentPanel } from './TemplateContentPanel'
+export { useQuickStartSettings } from './useQuickStartSettings'
